requirements.txt
# Fractal Agent Ecosystem - Python Dependencies
# Phase 0: Foundation

# LLM Providers
anthropic>=0.40.0
google-generativeai>=0.8.0

# DSPy Framework
dspy-ai>=2.5.0

# LangGraph Workflow
langgraph>=0.2.0
langchain>=0.3.0
langchain-core>=0.3.0

# Data & Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0

# Configuration
pyyaml>=6.0.0

# Utilities
tenacity>=8.0.0  # Retry logic
python-dateutil>=2.8.0

# Observability (Phase 5)
opentelemetry-api>=1.21.0
opentelemetry-sdk>=1.21.0
opentelemetry-exporter-otlp-proto-grpc>=1.21.0
psycopg2-binary>=2.9.0  # PostgreSQL for event store

# Testing (Phase 0)
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Development
black>=24.0.0  # Code formatting
ruff>=0.1.0  # Linting
mypy>=1.8.0  # Type checking

# Optional (Phase 1+)
# neo4j>=5.0.0  # GraphRAG - Phase 4
# qdrant-client>=1.7.0  # Vector store - Phase 4
# presidio-analyzer>=2.2.0  # PII detection - Phase 2
# presidio-anonymizer>=2.2.0  # PII redaction - Phase 2

# Knowledge & Context (Phase 6)
sentence-transformers>=2.2.0  # Embeddings for semantic search
numpy>=1.24.0  # Array operations for embeddings
# duckduckgo-search>=4.0.0  # Web search for context preparation (optional)
